docs(rule): adding a detailed description to the type input
Attempted to define a full type for the rule object.
After 800 lines of type definition, it became clear that this would need to become around six times longer to work.
Terraform can't reliably handle `any` types for elements within a list of typed objects.
Adding a speudo type in the description field instead.